feat(view-doc-source): add rum instrumentation
