+++ src/extension/module.js
@@ -11,6 +11,8 @@
  */
 /* eslint-disable no-console, no-alert */
 
+import { sampleRUM } from './rum.js';
+
 (() => {
   /**
    * @typedef {Object} ElemConfig
@@ -288,73 +290,6 @@
     DEFAULT: 0,
     CUSTOM: 1,
   };
-
-  /**
-   * Log RUM for sidekick telemetry.
-   * @private
-   * @param {string} checkpoint identifies the checkpoint in funnel
-   * @param {Object} data additional data for RUM sample
-   */
-  function sampleRUM(checkpoint, data = {}) {
-    sampleRUM.defer = sampleRUM.defer || [];
-    const defer = (fnname) => {
-      sampleRUM[fnname] = sampleRUM[fnname]
-        || ((...args) => sampleRUM.defer.push({ fnname, args }));
-    };
-    sampleRUM.drain = sampleRUM.drain
-      || ((dfnname, fn) => {
-        sampleRUM[dfnname] = fn;
-        sampleRUM.defer
-          .filter(({ fnname }) => dfnname === fnname)
-          .forEach(({ fnname, args }) => sampleRUM[fnname](...args));
-      });
-    sampleRUM.on = (chkpnt, fn) => {
-      sampleRUM.cases[chkpnt] = fn;
-    };
-    defer('observe');
-    defer('cw');
-    try {
-      window.hlx = window.hlx || {};
-      const sk = window.hlx.sidekick;
-      if (!window.hlx.rum) {
-        const usp = new URLSearchParams(sk.location.search);
-        const weight = (usp.get('hlx-sk-rum') === 'on') ? 1 : 10; // with parameter, weight is 1. Defaults to 10.
-        // eslint-disable-next-line no-bitwise
-        const hashCode = (s) => s.split('').reduce((a, b) => (((a << 5) - a) + b.charCodeAt(0)) | 0, 0);
-        const id = `${hashCode(sk.location.href)}-${new Date().getTime()}-${Math.random().toString(16).substr(2, 14)}`;
-        const random = Math.random();
-        const isSelected = (random * weight < 1);
-        // eslint-disable-next-line object-curly-newline
-        window.hlx.rum = { weight, id, random, isSelected, sampleRUM };
-      }
-      const { weight, id } = window.hlx.rum;
-      if (window.hlx && window.hlx.rum && window.hlx.rum.isSelected) {
-        const sendPing = () => {
-          // eslint-disable-next-line object-curly-newline, max-len, no-use-before-define
-          const body = JSON.stringify({ weight, id, referer: sk.location.href, generation: window.hlx.RUM_GENERATION, checkpoint, ...data });
-          const url = `https://rum.hlx.page/.rum/${weight}`;
-          // eslint-disable-next-line no-unused-expressions
-          navigator.sendBeacon(url, body);
-        };
-        sampleRUM.cases = sampleRUM.cases || {
-          cwv: () => sampleRUM.cwv(data) || true,
-          lazy: () => {
-            // use classic script to avoid CORS issues
-            const script = document.createElement('script');
-            script.src = 'https://rum.hlx.page/.rum/@adobe/helix-rum-enhancer@^1/src/index.js';
-            document.head.appendChild(script);
-            return true;
-          },
-        };
-        sendPing(data);
-        if (sampleRUM.cases[checkpoint]) {
-          sampleRUM.cases[checkpoint]();
-        }
-      }
-    } catch (error) {
-      // something went wrong
-    }
-  }
 
   /**
    * Retrieves project details from a host name.
@@ -3892,4 +3827,5 @@
 
   window.hlx = window.hlx || {};
   window.hlx.initSidekick = initSidekick;
+  window.hlx.sampleRUM = sampleRUM;
 })();
